继承抽象 Downloader
